Shift scripts/*.cs into BootstrapEnv project Enable auto-publish of packages when building them
