+++ BootstrapEnv/BootstrapEnv.csproj
@@ -5,9 +5,8 @@
     <Platforms>x64</Platforms>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="CommandLineParser" Version="2.7.82" />
-    <PackageReference Include="Improbable.Postgres" Version="0.0.3-preview" />
-    <PackageReference Include="Improbable.DatabaseSync" Version="0.0.2-preview" />
+    <PackageReference Include="McMaster.Extensions.CommandLineUtils" Version="2.5.0" />
+    <PackageReference Include="MedallionShell" Version="1.6.0" />
     <PackageReference Include="Serilog" Version="2.9.0" />
     <PackageReference Include="Serilog.Sinks.Console" Version="3.1.1" />
   </ItemGroup>
