+++ CSharpCodeGenerator/CSharpCodeGenerator/Program.cs
@@ -51,6 +51,8 @@ namespace CSharpCodeGenerator
             Log.Information("{Args}", string.Join(" ", ArgumentEscaper.EscapeAndConcatenate(args)));
 
             CommandLineApplication.Execute<Program>(args);
+
+            Log.CloseAndFlush();
         }
 
         private void OnExecute(CommandLineApplication app, CancellationToken token)
