+++ Workers.sln
@@ -16,6 +16,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Workers", "Workers", "{382C
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "CSharpCodeGenerator", "CSharpCodeGenerator", "{31E97862-E04A-4EB5-9EBA-AFDB0F453BA6}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BootstrapEnv", "BootstrapEnv\BootstrapEnv.csproj", "{B54DFDB5-4E63-4249-AAD9-AD652A815395}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -44,6 +46,12 @@ Global
 		{F1E2ECD6-1998-479F-AF15-D3B3131BAD5A}.Release|Any CPU.Build.0 = Release|Any CPU
 		{F1E2ECD6-1998-479F-AF15-D3B3131BAD5A}.Release|x64.ActiveCfg = Release|Any CPU
 		{F1E2ECD6-1998-479F-AF15-D3B3131BAD5A}.Release|x64.Build.0 = Release|Any CPU
+		{B54DFDB5-4E63-4249-AAD9-AD652A815395}.Debug|Any CPU.ActiveCfg = Debug|x64
+		{B54DFDB5-4E63-4249-AAD9-AD652A815395}.Debug|x64.ActiveCfg = Debug|x64
+		{B54DFDB5-4E63-4249-AAD9-AD652A815395}.Debug|x64.Build.0 = Debug|x64
+		{B54DFDB5-4E63-4249-AAD9-AD652A815395}.Release|Any CPU.ActiveCfg = Release|x64
+		{B54DFDB5-4E63-4249-AAD9-AD652A815395}.Release|x64.ActiveCfg = Release|x64
+		{B54DFDB5-4E63-4249-AAD9-AD652A815395}.Release|x64.Build.0 = Release|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
