--- scripts/launch.ps1
@@ -1,6 +1,0 @@
-$ErrorActionPreference = "Stop"
-
-cd "$PSScriptRoot/../"
-
-& spatial alpha local launch --main_config=./config/spatialos.json --launch_config=config/deployment.json
-return $?
